Fixing DTC not supported errors

+++ Raven.Tests/Transactions/Deletes.cs
@@ -32,6 +32,8 @@ namespace Raven.Tests.Transactions
 		[Fact]
 		public void DeletingDocumentInTransactionInNotVisibleBeforeCommit()
 		{
+            if (db.TransactionalStorage.SupportsDtc == false)
+                return;
 			db.Put("ayende", null, RavenJObject.Parse("{ayende:'oren'}"), new RavenJObject(), null);
             var transactionInformation = new TransactionInformation { Id = Guid.NewGuid().ToString(), Timeout = TimeSpan.FromMinutes(1) };
 			db.Delete("ayende", null, transactionInformation);
@@ -41,6 +43,8 @@ namespace Raven.Tests.Transactions
 		[Fact]
 		public void DeletingDocumentInTransactionInNotFoundInSameTransactionBeforeCommit()
 		{
+            if (db.TransactionalStorage.SupportsDtc == false)
+                return;
 			db.Put("ayende", null, RavenJObject.Parse("{ayende:'oren'}"), new RavenJObject(), null);
             var transactionInformation = new TransactionInformation { Id = Guid.NewGuid().ToString(), Timeout = TimeSpan.FromMinutes(1) };
 			db.Delete("ayende", null, transactionInformation);
@@ -51,6 +55,8 @@ namespace Raven.Tests.Transactions
 		[Fact]
 		public void DeletingDocumentAndThenAddingDocumentInSameTransactionCanWork()
 		{
+            if (db.TransactionalStorage.SupportsDtc == false)
+                return;
 			db.Put("ayende", null, RavenJObject.Parse("{ayende:'oren'}"), new RavenJObject(), null);
             var transactionInformation = new TransactionInformation { Id = Guid.NewGuid().ToString(), Timeout = TimeSpan.FromMinutes(1) };
 			db.Delete("ayende", null, transactionInformation);
@@ -65,6 +71,8 @@ namespace Raven.Tests.Transactions
 		[Fact]
 		public void DeletingDocumentInTransactionInRemovedAfterCommit()
 		{
+            if (db.TransactionalStorage.SupportsDtc == false)
+                return;
 			db.Put("ayende", null, RavenJObject.Parse("{ayende:'oren'}"), new RavenJObject(), null);
             var transactionInformation = new TransactionInformation { Id = Guid.NewGuid().ToString(), Timeout = TimeSpan.FromMinutes(1) };
 			db.Delete("ayende", null, transactionInformation);
